error refactor: collect as many type errors as possible
and report only at the very end in an ErrorSet. this patch causes the
checker to continue trying to typecheck after it has found an unbound
identifier (one where the type resolver returns nil). this means we
can still find type errors and other unbound identifiers and report a
more comprehensive and useful error at the end of the check.