perf fix: cache partial values and don't real-eval in partial
* the results of partial evaluation are now stored on an instance
variable on the node, so it does not have to be recomputed with a
recursive process
* we no longer drop into *real* evaluation of AND/OR nodes during
partial evaluation, saving another entire branch of recursion.

These two changes mean the algorithm is now linear. This can be verified
by adding a `puts` to the #value method of `AST::Or` and running
`PERF_SIZE=<your-input-size> rspec -t perf | wc -l`, and verifying the
number of times #value is called has a linear relationship to the input
size number. We used the same method to verify that the previous code
was superlinear (likely exponential).